Move Magic Home/flux_led coordinator to its own module
- Goal is to reduce conflicts in the future
